+++ src/components/NavbarClient.jsx
@@ -47,7 +47,7 @@ const NavbarClient = ({ loggedInUser, profileLink }) => {
               <>
                 <li>
                   <Link 
-                    href="/contact" 
+                    href="/aboutUs" 
                     className="block text-lg md:text-xl font-medium py-2 px-3 rounded-md transition-all hover:bg-gray-700 hover:text-gray-100"
                     onClick={() => setMenuOpen(false)}
                   >
